# server/models/Chytraskola_db/generated/UcitelModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ucitelModelGenerated.js PLEASE EDIT ../ucitelModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_Chytraskola_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * ucitel
      */
    const ucitelSchema = new mongoose.Schema({
      adresa: {
        type: "String"
      },
      datum_nastup: {
        type: "Date"
      },
      datum_odchod: {
        type: "Date"
      },
      email: {
        type: "String"
      },
      meno: {
        type: "String", 
        required: true
      },
      mobil: {
        type: "String"
      },
      pohlavie: {
        type: "String",
        enum: ["muz", "zena"], 
      },
      predmety: {
        type: "String"
      },
      priezvisko: {
        type: "String", 
        required: true
      },
      titul: {
        type: "String"
      },
      trieda: {
        type: "String"
      },
      zameranie: {
        type: "String"
      },
      // RELATIONS
      _predmety: [{
        type: Schema.ObjectId,
        ref: "predmety"
      }],
      _triedy: [{
        type: Schema.ObjectId,
        ref: "triedy"
      }],
      _uzivatel: {
        type: Schema.ObjectId,
        ref: "uzivatel"
      },
      
      // EXTERNAL RELATIONS
      /*
      _ucitel: {
        type: Schema.ObjectId,
        ref: "triedy"
      },
      _ucitel: [{
        type: Schema.ObjectId,
        ref: "predmety"
      }],
      _ucitel: [{
        type: Schema.ObjectId,
        ref: "skupina"
      }],
      _ucitel: [{
        type: Schema.ObjectId,
        ref: "rozvrh"
      }],
      _ucitel: [{
        type: Schema.ObjectId,
        ref: "dochadzka"
      }],
      */
    });

    generatedModel.setModel(db.connection.model("Ucitel", ucitelSchema));

    return ucitelSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS


  /**
  * ucitelModel.create
  *   @description CRUD ACTION create
  *
  */
  async create(item) {
    const obj = new generatedModel.model(item);
    return await obj.save();
  },
  
  /**
  * ucitelModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  async delete(id) {
    return await generatedModel.model.findByIdAndRemove(id);
  },
  
  /**
  * ucitelModel.findBy_triedy
  *   @description CRUD ACTION findBy_triedy
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_triedy(key) {
    return await generatedModel.model.find({ '_triedy' : key});
  },
  
  /**
  * ucitelModel.findBy_uzivatel
  *   @description CRUD ACTION findBy_uzivatel
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_uzivatel(key) {
    return await generatedModel.model.find({ '_uzivatel' : key});
  },
  
  /**
  * ucitelModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  async get(id) {
    return await generatedModel.model.findOne({ _id : id });
  },
  
  /**
  * ucitelModel.list
  *   @description CRUD ACTION list
  *
  */
  async list() {
    return await generatedModel.model.find();
  },
  
  /**
  * ucitelModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  async update(item) { 
    return await generatedModel.model.findOneAndUpdate({ _id: item._id }, item, {'new': true});
  },
  


};

export default generatedModel;