# server/swagger.yaml
openapi: 3.0.0
info:
  title: chytraskola
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /spravy_uzivatelia/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - spravy_uzivatelia
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - spravy_uzivatelia
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/spravy_uzivatelia"
    post:
      summary: CRUD ACTION update
      tags:
        - spravy_uzivatelia
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/spravy_uzivatelia"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/spravy_uzivatelia"
    
  #-----  student  -----
...
  /ucitel/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - ucitel
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - ucitel
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ucitel"
    post:
      summary: CRUD ACTION update
      tags:
        - ucitel
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ucitel"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ucitel"
  /ucitel/findBy_triedy/{key}:
    get:
      summary: CRUD ACTION findBy_triedy
      tags:
        - ucitel
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ucitel"
  /ucitel/findBy_uzivatel/{key}:
    get:
      summary: CRUD ACTION findBy_uzivatel
      tags:
        - ucitel
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ucitel"
    
  #-----  udalost  -----
...
  /uzivatel/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - uzivatel
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - uzivatel
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uzivatel"
    post:
      summary: CRUD ACTION update
      tags:
        - uzivatel
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uzivatel"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/uzivatel"
    
  #-----  vyucovanie_cas  -----
...
  /vyucovanie_cas/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - vyucovanie_cas
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - vyucovanie_cas
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vyucovanie_cas"
    post:
      summary: CRUD ACTION update
      tags:
        - vyucovanie_cas
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vyucovanie_cas"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/vyucovanie_cas"
    
  #-----  zadania  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    spravy_uzivatelia:
      properties:
        _id:
          type: string
        datum_cas:
          type: integer
        prijemca:
          type: string
        prilohy:
          type: object
        text:
          type: string
      required:
        -  _id
...
    ucitel:
      properties:
        _id:
          type: string
        adresa:
          type: string
        datum_nastup:
          type: integer
        datum_odchod:
          type: integer
        email:
          type: string
        meno:
          type: string
        mobil:
          type: string
        pohlavie:
          type: string
        predmety:
          type: string
        priezvisko:
          type: string
        titul:
          type: string
        trieda:
          type: string
        zameranie:
          type: string
      required:
        -  _id
        - meno
        - priezvisko
...
    uzivatel:
      properties:
        _id:
          type: string
        datum_aktualizacia:
          type: integer
        datum_vznik:
          type: integer
        heslo:
          type: string
        meno:
          type: string
        rola:
          type: string
        status:
          type: string
      required:
        -  _id
        - heslo
        - meno
    vyucovanie_cas:
      properties:
        _id:
          type: string
        cas_od_do:
          type: string
        poradie_hodiny:
          type: number
      required:
        -  _id
...